progress on fatigue functionality
